refactor: extract shared CI helpers, remove push trigger
Move duplicated setOutput and GitHub API client into ci-helpers.mjs
so all three CI scripts import from a single source of truth.
Remove the temporary push trigger and its event handling from the
workflow, leaving only issue_comment and workflow_dispatch triggers.

Made-with: Cursor
